Made a proper parser class, cleaned, & commented

+++ StandardCalculator/src/data/expression/binaryExpressions/BinaryExpression.java
@@ -7,7 +7,7 @@ public abstract class BinaryExpression extends NonTerminalExpression {
     protected Expression y;
 
     //abstract public Boolean checkSymbol();
-    abstract public float solve();
+    abstract public double solve();
 
     public void setX(Expression expression) {
         this.x = expression;
